requirements.txt
fastapi
uvicorn
pdfplumber
openai
nltk
scikit-learn
numpy
pydantic
python-multipart
aiofiles
gradio
asyncio
langdetect
typing
pydantic-settings
spacy
subprocess
sys
python-magic
io
docx
